Refactor event listener functions
I changed the event listener functions so that they use one
reusable function.

With this change, I learned how to better utilize "this"

At first, I tried making the new changeMode function as the
function that is called. However, because I'm using "this" as the
parameter I have to wrap it in an anonymous function.

I did not know this initially and had to research why nothing was
working.